part4, 4.23 fix tests

+++ part4_blog/tests/api-blogs.test.js
@@ -1,37 +1,68 @@
-const { test, after, beforeEach, describe } = require('node:test')
+const { test, after, beforeEach, before , describe } = require('node:test')
 const assert = require('node:assert/strict')
 const mongoose = require('mongoose')
 const supertest = require('supertest')
 const app = require('../app')
 const api = supertest(app)
 const Blog = require('../models/blog')
+const User = require('../models/user')
 const db = require('../database/mongodb')
 
-const blogs = [
-  { 'title': 'A New Level',
+let userObj = {
+  user : 'John Doe',
+  username : 'jdoe',
+  password : '123456*',
+  blogs : []
+}
+
+let blogsObj = [
+  {
+    'title': 'A New Level',
     'author': 'Dimebag Darrell',
     'url': 'https://es.wikipedia.org/wiki/Pantera_(banda)',
-    'likes': 666
+    'likes': 666,
+    'user': null
   }
 ]
 
+let userId
+let savedBlogs
+let token
+
 db.connect()
+
+before(async () => {
+  await User.deleteMany({})
+  const userCreatedResponse = await api.post('/api/users').send(userObj)
+  userId = userCreatedResponse.body.id
+
+  //login
+  const loginResponse = await api.post('/api/login').send({ username : userObj.username, password : userObj.password })
+  token = `Bearer ${loginResponse.body.token}`
+})
 
 beforeEach(async () => {
   await Blog.deleteMany({})
-  await Blog.insertMany(blogs)
+
+  const blogs = blogsObj.map( b => { return { ...b, user: userId }})
+
+  savedBlogs = await Blog.insertMany(blogs)
+  const userModel = await User.findById(userId)
+
+  userModel.blogs = savedBlogs.map( b => b.id )
+  await userModel.save()
 })
 
 describe('when a collection of blogs is retrieved...', () => {
   test('blogs are a list of json with one element (test database)', async () => {
-    const response = await api.get('/api/blogs')
+    const response = await api.get('/api/blogs').set('Authorization', token)
       .expect('Content-Type', /application\/json/)
 
     assert.strictEqual(response.body.length, 1)
   })
 
   test('blogs use id as a key instead _id', async () => {
-    const response = await api.get('/api/blogs')
+    const response = await api.get('/api/blogs').set('Authorization', token)
       .expect(200)
 
     assert(Object.keys(response.body[0]).includes('id'))
@@ -47,16 +78,16 @@ describe('when a blog is creted...', () => {
       likes : 1
     }
 
-    const createBlogResponse = await api.post('/api/blogs')
+    const createBlogResponse = await api.post('/api/blogs').set('Authorization', token)
       .send(newBlog)
       .expect(201)
 
     assert.partialDeepStrictEqual(createBlogResponse.body, newBlog)
 
-    const getBlogsResponse = await api.get('/api/blogs')
+    const getBlogsResponse = await api.get('/api/blogs').set('Authorization', token)
       .expect(200)
 
-    assert.strictEqual(getBlogsResponse.body.length - blogs.length, 1)
+    assert.strictEqual(getBlogsResponse.body.length - savedBlogs.length, 1)
   })
 
   test('blog\'s likes are defaulted to zero when are missing from request', async () => {
@@ -66,7 +97,7 @@ describe('when a blog is creted...', () => {
       url : 'https://en.wikipedia.org/wiki/Rust_in_Peace',
     }
 
-    const createBlogResponse = await api.post('/api/blogs')
+    const createBlogResponse = await api.post('/api/blogs').set('Authorization', token)
       .send(newBlog)
       .expect(201)
 
@@ -79,7 +110,7 @@ describe('when a blog is creted...', () => {
       url : 'https://en.wikipedia.org/wiki/Rust_in_Peace',
       likes : 0
     }
-    await api.post('/api/blogs')
+    await api.post('/api/blogs').set('Authorization', token)
       .send(newBlog)
       .expect(400)
   })
@@ -91,7 +122,7 @@ describe('when a blog is creted...', () => {
       likes : 0
     }
 
-    await api.post('/api/blogs')
+    await api.post('/api/blogs').set('Authorization', token)
       .send(newBlog)
       .expect(400)
   })
@@ -100,10 +131,10 @@ describe('when a blog is creted...', () => {
 describe('when a blog is deleted...', () => {
   test('with a correct id, the blog is succesfully deleted and 204 is returned', async () => {
 
-    const blogs = await Blog.find({})
-    const blogId = blogs[0].id
+    const blog = await Blog.findOne({ user : userId })
+    const blogId = blog.id
 
-    await api.delete(`/api/blogs/${blogId}`).expect(204)
+    await api.delete(`/api/blogs/${blogId}`).set('Authorization', token).expect(204)
 
     const found = await Blog.findById(blogId)
     assert(!found)
@@ -114,14 +145,16 @@ describe('when a blog is deleted...', () => {
 describe('when a blog is updated...', () => {
   test('with a correct id and only with likes in the request body, blog is successfully updated and 200 is returned', async () => {
 
-    const blog = (await Blog.find({}))[0]
+    const blog = await Blog.findOne({ user : userId })
     const blogId = blog.id
 
     const bodyRequest = { likes : 999 }
 
-    const updatedBlog = await api.put(`/api/blogs/${blogId}`).send(bodyRequest).expect(200)
+    const updatedBlog = await api.put(`/api/blogs/${blogId}`).set('Authorization', token).send(bodyRequest).expect(200)
 
     const expectedBlog = { ...blog.toJSON() , ...bodyRequest }
+    expectedBlog.user = expectedBlog.user.toString()
+
     assert.deepStrictEqual(updatedBlog.body, expectedBlog)
   })
 })
